+++ lib/plugins/index.js
@@ -10,7 +10,7 @@ module.exports = [
   require('./invoke.js'),
   require('./info.js'),
   require('./logs.js'),
-  require('./metrics/metrics.js'),
+  require('./metrics.js'),
   require('./print/print.js'),
   require('./remove/remove.js'),
   require('./rollback/index.js'),
