+++ lib/plugins/metrics.js
